% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalkr.R
\docType{package}
\name{crosswalkr}
\alias{crosswalkr}
\alias{crosswalkr-package}
\title{crosswalkr: Rename and Encode Data Frames Using External Crosswalk Files}
\description{
A pair of functions for renaming and encoding data frames using
external crosswalk files. Especially useful when constructing
master data sets from multiple smaller data sets that do not name
or encode variables consistently across files. Based on
-renamefrom- and -encodefrom- Stata commands.
}
