% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renamefrom.R
\name{renamefrom_}
\alias{renamefrom_}
\title{Standard evaluation version of \code{\link{renamefrom}}}
\usage{
renamefrom_(.data, cw_file, raw, clean, label = NULL, delimiter = NULL,
  sheet = NULL, drop_extra = TRUE, case_ignore = TRUE,
  keep_label = FALSE, name_label = FALSE)
}
\arguments{
\item{.data}{Data frame}

\item{cw_file}{External crosswalk file with columns representing
\code{raw} (current) column names, \code{clean} (new) column
names, and labels (optional). Values in \code{raw} and
\code{clean} columns must be unique (1:1 match) or an error
will be thrown. Acceptable file types include: delimited (.csv,
.tsv, or other), R (.rda, .rdata, .rds), or Stata (.dta).}

\item{raw}{Name of column in \code{cw_file} that contains column
names of current data frame.}

\item{clean}{Name of column in \code{cw_file} that contains new
column names.}

\item{label}{Name of column in \code{cw_file} with labels for
columns (default = \code{NULL}).}

\item{delimiter}{String delimiter used to parse
\code{cw_file}. Only necessary if using a delimited file that
isn't a comma-separated or tab-separated file (guessed by
function based on file ending) (default = \code{NULL}).}

\item{sheet}{Specify sheet if \code{cw_file} is an Excel file and
required sheet isn't the first one (default = \code{NULL}).}

\item{drop_extra}{Drop extra columns in current data frame if they
are not matched in \code{cw_file} (default = \code{NULL}).}

\item{case_ignore}{Ignore case when matching current (\code{raw})
column names with new (\code{clean}) column names (default =
\code{TRUE}).}

\item{keep_label}{Keep current label, if any, on data frame columns
that aren't matched in \code{cw_file}. Default \code{FALSE}
means that unmatched columns have any existing labels set to
\code{NULL}.}

\item{name_label}{Use old (\code{raw}) column name as new
(\code{clean}) column name label. Cannot be used if
\code{label} option is set.}
}
\description{
Standard evaluation version of \code{\link{renamefrom}}
}
