% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stcrosswalk}
\alias{stcrosswalk}
\title{State crosswalk data set.}
\format{
A data frame with 51 rows and 7 variables:
\describe{
  \item{stfips}{Two-digit state FIPS codes}
  \item{stabbr}{Two-letter state abbreviation}
  \item{stname}{Full state name}
  \item{cenreg}{Census region number}
  \item{cenregnm}{Census region name}
  \item{cendiv}{Census division number}
  \item{cendivnm}{Census division name}
}
}
\usage{
stcrosswalk
}
\description{
An example state crosswalk. Includes information for all states
plus the District of Columbia.
}
\keyword{datasets}
